ubung2

+++ src/exercise2/PiApproximation.java
@@ -4,7 +4,7 @@ public class PiApproximation{
 
   public static void main(String[] args){
     
-     float pi = 4 * (1 - ((1/3) + (1/5)) - ((1/7) + (1/11)) - ((1/13) + (1/17)) - ((1/19) + (1/23)));
+     double pi = 4.0 * ( 1.0 - (1.0/3.0) + (1.0/5.0) - (1.0/7.0) + (1.0/11.0) - (1.0/13.0) + (1.0/17.0) - (1.0/19.0) + (1.0/23.0) );
 
      System.out.println("Pi: " + pi); //how to converge to 3.14?
   }
